Add multiple houses handling

+++ src/srv/routes/homes.ts
@@ -20,9 +20,12 @@ type UserHomeTokenCreateParams = {
 }
 
 type UserHomeTokenDeleteParams = {
-  userId: string,
   homeId: string,
   tokenId: string
+}
+
+type UserHomeDeleteParams = {
+  homeId: string
 }
 
 type HomeToken = {
@@ -34,11 +37,17 @@ type HomeToken = {
 
 type HomeTokenCreateBody = Omit<HomeToken, '_id' | 'homeId'>;
 
-type UserHomeDbResponse = {
+type UserHomeType = {
   _id: string,
   name: string,
   userId: string
   tokens: HomeToken[]
+}
+
+type UserHomeCreateBody = Omit<UserHomeType, '_id'>;
+
+type HomesQueryString = { 
+  tokenData: string | undefined 
 }
 
 export const homesRoutes = async (fastify: FastifyInstance, options: RouteOptions) => {
@@ -48,10 +57,12 @@ export const homesRoutes = async (fastify: FastifyInstance, options: RouteOption
     onRequest: [fastify.authenticate]
   }, async (request, reply) => {
     try {
-      const userId = request.user.userId;
+      const { userId } = request.user;
+      const { tokenData } = await request.query as HomesQueryString;
 
       const db = fastify.mongo.client.db(config.dbName);
-      const homesData = await db.collection('homesWithTokens').find({
+      const collection = db.collection(tokenData ? 'homesWithTokens' : 'homes'); 
+      const homesData = await collection.find({
         userId: new ObjectId(userId)
       }).toArray();
 
@@ -61,14 +72,76 @@ export const homesRoutes = async (fastify: FastifyInstance, options: RouteOption
     }
   })
 
+  fastify.post('/api/homes', {
+    onRequest: [fastify.authenticate]
+  }, async (request, reply) => {
+    try {
+      const { userId } = request.user;
+      const { name } = request.body as UserHomeCreateBody;
+
+      const db = fastify.mongo.client.db(config.dbName); 
+      await db.collection('homes').insertOne({
+        userId: new ObjectId(userId),
+        tokens: [],
+        name
+      });
+
+      return reply.send({ status: 'CREATED' });
+    } catch (error) {
+      fastify.log.error(error);
+    }
+  })
+
+  fastify.delete('/api/homes/:homeId', {
+    onRequest: [fastify.authenticate]
+  }, async (request, reply) => {
+    try {
+      const { homeId } = request.params as UserHomeDeleteParams;
+      const { userId } = request.user;
+
+      const db = fastify.mongo.client.db(config.dbName);
+
+      const userHomes = await db.collection('homes').find({
+        userId: new ObjectId(userId)
+      }).toArray();
+      const userIsOwnerOfHome = userHomes.find((home) => home._id.toString() === homeId)
+
+      if (!userIsOwnerOfHome) {
+        return reply.status(401).send({ error: 'Unauthorized' });
+      }
+
+      await db.collection('tokens').deleteMany(
+        { 'homeId': new ObjectId(homeId) }, 
+      );
+
+      await db.collection('homes').deleteOne(
+        { '_id': new ObjectId(homeId) }, 
+      );
+
+      return reply.send({status: 'DELETED'});
+    } catch (error) {
+      fastify.log.error(error);
+    }
+  })
+
   fastify.post('/api/homes/:homeId/token', {
-    onRequest: [fastify.authenticate, fastify.checkHomePermissions]
+    onRequest: [fastify.authenticate]
   }, async (request, reply) => {
     try {
       const { homeId } = request.params as UserHomeTokenCreateParams;
       const { name, value } = request.body as HomeTokenCreateBody;
+      const { userId } = request.user;
 
       const db = fastify.mongo.client.db(config.dbName);
+
+      const userHomes = await db.collection('homes').find({
+        userId: new ObjectId(userId)
+      }).toArray();
+      const userIsOwnerOfHome = userHomes.find((home) => home._id.toString() === homeId)
+
+      if (!userIsOwnerOfHome) {
+        return reply.status(401).send({ error: 'Unauthorized' });
+      }
 
       await db.collection('tokens').insertOne({
         name,
@@ -92,12 +165,22 @@ export const homesRoutes = async (fastify: FastifyInstance, options: RouteOption
   })
 
   fastify.delete('/api/homes/:homeId/token/:tokenId', {
-    onRequest: [fastify.authenticate, fastify.checkHomePermissions]
+    onRequest: [fastify.authenticate]
   }, async (request, reply) => {
     try {
       const { homeId, tokenId } = request.params as UserHomeTokenDeleteParams;
+      const { userId } = request.user;
 
       const db = fastify.mongo.client.db(config.dbName);
+
+      const userHomes = await db.collection('homes').find({
+        userId: new ObjectId(userId)
+      }).toArray();
+      const userIsOwnerOfHome = userHomes.find((home) => home._id.toString() === homeId)
+
+      if (!userIsOwnerOfHome) {
+        return reply.status(401).send({ error: 'Unauthorized' });
+      }
 
       const deletedData = await db.collection('tokens').deleteOne(
         { '_id': new ObjectId(tokenId) },
